Implement Exporter for all MarkupExporters
